requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1
django-phonenumber-field==6.4.0
idna==3.3
phonenumbers==8.12.54
phonenumberslite==8.12.54
pycparser==2.21
PyJWT==2.5.0
PyNaCl==1.5.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
